Restructured the files
