+++ app/src/main/java/com/example/pricingpal/view/homepage/guest/OwnerRegisteration.kt
@@ -51,6 +51,10 @@ import com.example.pricingpal.ui.theme.Anti_flash_white
 import com.example.pricingpal.ui.theme.Cornflower_blue
 import com.example.pricingpal.ui.theme.Periwinkle
 import com.example.pricingpal.ui.theme.Persian_indigo
+import com.example.pricingpal.view.repetitivefunctions.arrowNavigationBar
+import com.example.pricingpal.view.repetitivefunctions.emailInput
+import com.example.pricingpal.view.repetitivefunctions.innerPricingBar
+import com.example.pricingpal.view.repetitivefunctions.passwordInput
 
 @Composable
 fun ownerRegistration(){
@@ -66,21 +70,8 @@ fun ownerRegistration(){
         colors = CardDefaults.cardColors(containerColor = Periwinkle)
     ) {
 
-        Row(modifier = Modifier
-            .border(4.dp, color = Persian_indigo)
-            .fillMaxWidth()
-            .background(color = Cornflower_blue, shape = RectangleShape),
-            horizontalArrangement = Arrangement.Center,
-        ) {
-            Image(
-                //Imports image from resource folder
-                painter = painterResource(id = R.drawable.logo),
-                //description of the image for accessibility
-                contentDescription = "Pictures of paws",
-                modifier = Modifier
-                    .padding(15.dp)
-            )
-        }
+        arrowNavigationBar()
+        innerPricingBar()
 
         Column(
             modifier = Modifier
@@ -100,9 +91,9 @@ fun ownerRegistration(){
             Spacer(modifier = Modifier.height(25.dp))
             ownerOrganizationInput()
             Spacer(modifier = Modifier.height(25.dp))
-            ownerEmailInput()
+            emailInput()
             Spacer(modifier = Modifier.height(25.dp))
-            ownerPasswordInput()
+            passwordInput()
             Spacer(modifier = Modifier.height(25.dp))
 
             ElevatedButton(
@@ -126,27 +117,6 @@ fun ownerRegistration(){
             }
 
             Spacer(modifier = Modifier.height(60.dp))
-
-            ElevatedButton(
-                onClick = { /*TODO*/ },
-                shape = RectangleShape,
-                colors = ButtonDefaults.buttonColors(Cornflower_blue),
-                elevation = ButtonDefaults.buttonElevation(8.dp),
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(90.dp)
-                    .padding(start = 70.dp, top = 15.dp, end = 70.dp, bottom = 15.dp)
-                    .border(4.dp, color = Persian_indigo),
-
-                ) {
-                Text(
-                    textAlign = TextAlign.Center,
-                    text = "Home",
-                    fontSize = 30.sp,
-                    color = Color.Black,
-                )
-
-            }
         }
     }
 }
@@ -185,101 +155,4 @@ fun ownerOrganizationInput(){
             color = Color.DarkGray
         )
     }
-}
-
-@Composable
-fun ownerEmailInput(){
-    var ownerEmail by remember { mutableStateOf("") }
-    TextField(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(60.dp)
-            .padding(start = 30.dp, end = 30.dp),
-        value = ownerEmail,
-        onValueChange = {ownerEmail = it},
-        textStyle = TextStyle.Default.copy(fontSize = 20.sp) ,
-        placeholder = { Text("Enter email", fontSize = 20.sp) },
-        supportingText = { Text(text = "*required") },
-        leadingIcon = { Icon(imageVector = Icons.Filled.Email, contentDescription = "Email Icon") },
-        colors = TextFieldDefaults.colors(
-            focusedContainerColor = Anti_flash_white,
-            unfocusedContainerColor = Anti_flash_white,
-            unfocusedIndicatorColor = Anti_flash_white,
-            focusedIndicatorColor = Persian_indigo
-        ),
-        shape = RectangleShape,
-    )
-    /** I did this in replacement of the supporting text*/
-    Row(horizontalArrangement = Arrangement.Start,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 5.dp, start = 50.dp)
-    ) {
-        Text(
-            text = "*required",
-            fontSize = 20.sp,
-            color = Color.DarkGray
-        )
-    }
-}
-
-@Composable
-fun ownerPasswordInput(){
-    var ownerPassword by remember { mutableStateOf("") }
-    TextField(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(60.dp)
-            .padding(start = 30.dp, end = 30.dp),
-        value = ownerPassword,
-        onValueChange = { ownerPassword = it },
-        textStyle = TextStyle.Default.copy(fontSize = 20.sp),
-        placeholder = { Text("Enter password", fontSize = 20.sp) },
-        /** The support text will not work if you have a modifier.*/
-        /** The support text will not work if you have a modifier.*/
-        //supportingText = { Text(text = "*required")},
-        visualTransformation = PasswordVisualTransformation(),
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
-        leadingIcon = { Icon(imageVector = Icons.Filled.Lock, contentDescription = "Lock Icon") },
-        trailingIcon = {
-            IconButton(onClick = { /*TODO*/ }) {
-                Icon(imageVector = ImageVector.vectorResource(id = R.drawable.eye), contentDescription = "Lock Icon")
-            }
-        },
-        /** This is for the hidden Icon that will turn the password hidden again.
-
-        trailingIcon = {
-        IconButton(onClick = { /*TODO*/ }) {
-        Icon(imageVector = ImageVector.vectorResource(id = R.drawable.eye), contentDescription = "Lock Icon")}
-        },
-         */
-        /** This is for the hidden Icon that will turn the password hidden again.
-
-        trailingIcon = {
-        IconButton(onClick = { /*TODO*/ }) {
-        Icon(imageVector = ImageVector.vectorResource(id = R.drawable.eye), contentDescription = "Lock Icon")}
-        },
-         */
-        supportingText = { Text(text = "*required") },
-        colors = TextFieldDefaults.colors(
-            focusedContainerColor = Anti_flash_white,
-            unfocusedContainerColor = Anti_flash_white,
-            unfocusedIndicatorColor = Anti_flash_white,
-            focusedIndicatorColor = Persian_indigo
-        ),
-        shape = RectangleShape,
-    )
-    /** I did this in replacement of the supporting text*/
-    Row(horizontalArrangement = Arrangement.Start,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 5.dp, start = 50.dp)
-    ) {
-        Text(
-            text = "*required",
-            fontSize = 20.sp,
-            color = Color.DarkGray
-        )
-    }
-
 }
