+++ app/src/main/java/com/example/pricingpal/view/homepage/login/CreatePassword.kt
@@ -53,6 +53,9 @@ import com.example.pricingpal.ui.theme.Anti_flash_white
 import com.example.pricingpal.ui.theme.Cornflower_blue
 import com.example.pricingpal.ui.theme.Periwinkle
 import com.example.pricingpal.ui.theme.Persian_indigo
+import com.example.pricingpal.view.repetitivefunctions.arrowNavigationBar
+import com.example.pricingpal.view.repetitivefunctions.innerPricingBar
+import com.example.pricingpal.view.repetitivefunctions.passwordInput
 
 @Composable
 fun createPassword(){
@@ -68,21 +71,8 @@ fun createPassword(){
         colors = CardDefaults.cardColors(containerColor = Periwinkle)
     ) {
 
-        Row(modifier = Modifier
-            .border(4.dp, color = Persian_indigo)
-            .fillMaxWidth()
-            .background(color = Cornflower_blue, shape = RectangleShape),
-            horizontalArrangement = Arrangement.Center,
-        ) {
-            Image(
-                //Imports image from resource folder
-                painter = painterResource(id = R.drawable.logo),
-                //description of the image for accessibility
-                contentDescription = "Pictures of paws",
-                modifier = Modifier
-                    .padding(15.dp)
-            )
-        }
+        arrowNavigationBar()
+        innerPricingBar()
 
         Column(
             modifier = Modifier
@@ -109,7 +99,7 @@ fun createPassword(){
             )
 
             Spacer(modifier = Modifier.height(25.dp))
-            newPasswordInput()
+            passwordInput()
             Spacer(modifier = Modifier.height(25.dp))
             confirmPasswordInput()
             Spacer(modifier = Modifier.height(35.dp))
@@ -135,92 +125,11 @@ fun createPassword(){
             }
 
             Spacer(modifier = Modifier.height(90.dp))
-
-            ElevatedButton(
-                onClick = { /*TODO*/ },
-                shape = RectangleShape,
-                colors = ButtonDefaults.buttonColors(Cornflower_blue),
-                elevation = ButtonDefaults.buttonElevation(8.dp),
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(90.dp)
-                    .padding(start = 70.dp, top = 15.dp, end = 70.dp, bottom = 15.dp)
-                    .border(4.dp, color = Persian_indigo),
-
-                ) {
-                Text(
-                    textAlign = TextAlign.Center,
-                    text = "Cancel",
-                    fontSize = 30.sp,
-                    color = Color.Black,
-                )
-
-            }
         }
     }
 }
 
 
-
-@Composable
-fun newPasswordInput(){
-    var newPassword by remember { mutableStateOf("") }
-    TextField(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(60.dp)
-            .padding(start = 30.dp, end = 30.dp),
-        value = newPassword,
-        onValueChange = { newPassword = it },
-        textStyle = TextStyle.Default.copy(fontSize = 20.sp),
-        placeholder = { Text("Enter new password", fontSize = 20.sp) },
-        /** The support text will not work if you have a modifier.*/
-        /** The support text will not work if you have a modifier.*/
-        //supportingText = { Text(text = "*required")},
-        visualTransformation = PasswordVisualTransformation(),
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
-        leadingIcon = { Icon(imageVector = Icons.Filled.Lock, contentDescription = "Lock Icon") },
-        trailingIcon = {
-            IconButton(onClick = { /*TODO*/ }) {
-                Icon(imageVector = ImageVector.vectorResource(id = R.drawable.eye), contentDescription = "Lock Icon")
-            }
-        },
-        /** This is for the hidden Icon that will turn the password hidden again.
-
-        trailingIcon = {
-        IconButton(onClick = { /*TODO*/ }) {
-        Icon(imageVector = ImageVector.vectorResource(id = R.drawable.eye), contentDescription = "Lock Icon")}
-        },
-         */
-        /** This is for the hidden Icon that will turn the password hidden again.
-
-        trailingIcon = {
-        IconButton(onClick = { /*TODO*/ }) {
-        Icon(imageVector = ImageVector.vectorResource(id = R.drawable.eye), contentDescription = "Lock Icon")}
-        },
-         */
-        supportingText = { Text(text = "*required") },
-        colors = TextFieldDefaults.colors(
-            focusedContainerColor = Anti_flash_white,
-            unfocusedContainerColor = Anti_flash_white,
-            unfocusedIndicatorColor = Anti_flash_white,
-            focusedIndicatorColor = Persian_indigo
-        ),
-        shape = RectangleShape,
-    )
-    /** I did this in replacement of the supporting text*/
-    Row(horizontalArrangement = Arrangement.Start,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 5.dp, start = 50.dp)
-    ) {
-        Text(
-            text = "*required",
-            fontSize = 20.sp,
-            color = Color.DarkGray
-        )
-    }
-}
 
 @Composable
 fun confirmPasswordInput(){
@@ -235,7 +144,6 @@ fun confirmPasswordInput(){
         textStyle = TextStyle.Default.copy(fontSize = 20.sp),
         placeholder = { Text("Enter confirm new password", fontSize = 20.sp) },
         /** The support text will not work if you have a modifier.*/
-        /** The support text will not work if you have a modifier.*/
         //supportingText = { Text(text = "*required")},
         visualTransformation = PasswordVisualTransformation(),
         keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
@@ -245,13 +153,6 @@ fun confirmPasswordInput(){
                 Icon(imageVector = ImageVector.vectorResource(id = R.drawable.eye), contentDescription = "Lock Icon")
             }
         },
-        /** This is for the hidden Icon that will turn the password hidden again.
-
-        trailingIcon = {
-        IconButton(onClick = { /*TODO*/ }) {
-        Icon(imageVector = ImageVector.vectorResource(id = R.drawable.eye), contentDescription = "Lock Icon")}
-        },
-         */
         /** This is for the hidden Icon that will turn the password hidden again.
 
         trailingIcon = {
