+++ app/src/main/java/com/example/pricingpal/view/homepage/volunteer/VolunteerCategoryList.kt
@@ -1,7 +1,8 @@
-package com.example.pricingpal.view.homepage.guest
+package com.example.pricingpal.view.homepage.volunteer
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
@@ -11,31 +12,35 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Divider
+import androidx.compose.material3.ElevatedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.example.pricingpal.R
 import com.example.pricingpal.ui.theme.Cornflower_blue
-import com.example.pricingpal.ui.theme.Periwinkle
 import com.example.pricingpal.ui.theme.Persian_indigo
 import com.example.pricingpal.view.repetitivefunctions.arrowNavigationBar
 import com.example.pricingpal.view.repetitivefunctions.pricingPalBar
 import com.example.pricingpal.view.repetitivefunctions.searchBar
 
 @Composable
-fun companiesTitle(){
+fun volunteerCategory(){
     Card(
         shape = RectangleShape,
         modifier = Modifier
-            .padding(top = 15.dp)
             .padding(15.dp),
         elevation = CardDefaults.cardElevation(12.dp),
     ) {
@@ -50,71 +55,69 @@ fun companiesTitle(){
                     shape = RectangleShape
                 )
                 .fillMaxWidth()
-                .padding(15.dp)
+                //.padding(15.dp)
         )
         {
-
-            Text(
-                //shows the name of the category
-                text = "Organizations",
-                fontSize = 60.sp,
-                // have it in bold to make it stand out
-                fontWeight = FontWeight.Bold,
-                color = Color.Black,
+            Image(
+                painter = painterResource(id =R.drawable.rectangle_22),
+                contentDescription = "Accessories image",
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(220.dp),
+                contentScale = ContentScale.Crop,
+                alpha = 0.8F
             )
 
-            Text(
-                //shows the name of the category
-                text = "Select only one organization",
-                fontSize = 30.sp,
-                color = Color.Black,
-            )
+            Row(modifier = Modifier
+                .border(4.dp, color = Persian_indigo)
+                .fillMaxWidth()
+                .height(60.dp)
+                .background(color = Cornflower_blue, shape = RectangleShape),
+                horizontalArrangement = Arrangement.Center,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+
+                Text(
+                    //shows the name of the category
+                    text = "Category",
+                    fontSize = 30.sp,
+                    // have it in bold to make it stand out
+                    fontWeight = FontWeight.Bold,
+                    color = Color.Black,
+                )
+
+            }
         }
     }
 }
 
 @Composable
-fun companyName(){
-    Card(
+fun volunteerHome(){
+    ElevatedButton(
+        onClick = { /*TODO*/ },
+        shape = RectangleShape,
+        colors = ButtonDefaults.buttonColors(Cornflower_blue),
+        elevation = ButtonDefaults.buttonElevation(8.dp),
         modifier = Modifier
-            // padding around the card
-            .padding(15.dp)
-            .border(
-                // puts a border around the card
-                border = BorderStroke(4.dp, color = Persian_indigo),
-                // shapes the card
-                shape = RectangleShape
-            ),
-        // puts a shadow under the card to make it pop out
-        elevation = CardDefaults.cardElevation(12.dp),
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-            modifier = Modifier
-                // changes the color of the card
-                .background(color = Periwinkle, shape = RectangleShape)
-                // changes the size of the card
-                .fillMaxWidth()
-                .height(80.dp)
-        )
-        {
-            Text(
-                text = "Organization Name",
-                fontSize = 30.sp,
-                color = Color.Black,
-                modifier = Modifier
-                    .padding(start = 20.dp)
-                    .align(alignment = Alignment.CenterVertically)
-            )
+            .fillMaxWidth()
+            .height(90.dp)
+            .padding(start = 70.dp, top = 15.dp, end = 70.dp, bottom = 15.dp)
+            .border(4.dp, color = Persian_indigo),
 
-        }
+        ) {
+        Text(
+            textAlign = TextAlign.Center,
+            text = "Home",
+            fontSize = 30.sp,
+            color = Color.Black,
+        )
+
     }
 }
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-fun guestCompanyList(){
+fun volunteerCategoryList(){
     arrowNavigationBar()
     LazyColumn(
         modifier = Modifier
@@ -127,15 +130,13 @@ fun guestCompanyList(){
             searchBar()
             Divider(thickness = 4.dp, color = Persian_indigo)
         }
-        item {
-            companiesTitle()
-        }
-        for(i in 1..10) {
+        for(i in 1..2) {
             item {
-                companyName()
+                volunteerCategory()
             }
+        }
+        item{
+            volunteerHome()
         }
     }
 }
-
-
