+++ app/src/main/java/com/example/pricingpal/view/repetitivefunctions/Buttons.kt
@@ -1,0 +1,125 @@
+package com.example.pricingpal.view.repetitivefunctions
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.ElevatedButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.RectangleShape
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.pricingpal.ui.theme.Cornflower_blue
+import com.example.pricingpal.ui.theme.Persian_indigo
+
+@Composable
+fun homeButton(){
+    ElevatedButton(
+        onClick = { /*TODO*/ },
+        shape = RectangleShape,
+        colors = ButtonDefaults.buttonColors(Cornflower_blue),
+        elevation = ButtonDefaults.buttonElevation(8.dp),
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(90.dp)
+            .padding(start = 70.dp, top = 15.dp, end = 70.dp, bottom = 15.dp)
+            .border(4.dp, color = Persian_indigo),
+
+        ) {
+        Text(
+            textAlign = TextAlign.Center,
+            text = "Home",
+            fontSize = 30.sp,
+            color = Color.Black,
+        )
+
+    }
+}
+
+@Composable
+fun deleteCancelButton(){
+    ElevatedButton(
+        onClick = { /*TODO*/ },
+        shape = RectangleShape,
+        colors = ButtonDefaults.buttonColors(Cornflower_blue),
+        elevation = ButtonDefaults.buttonElevation(8.dp),
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(90.dp)
+            .padding(start = 70.dp, top = 15.dp, end = 70.dp, bottom = 15.dp)
+            .border(4.dp, color = Persian_indigo),
+
+        ) {
+        Text(
+            textAlign = TextAlign.Center,
+            text = "Cancel",
+            fontSize = 30.sp,
+            color = Color.Black,
+        )
+
+    }
+}
+
+@Composable
+fun signUpButton(){
+    ElevatedButton(
+        onClick = { /*TODO*/ },
+        shape = RectangleShape,
+        colors = ButtonDefaults.buttonColors(Cornflower_blue),
+        elevation = ButtonDefaults.buttonElevation(8.dp),
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(120.dp)
+            .padding(start = 25.dp, top = 15.dp, end = 25.dp, bottom = 15.dp)
+            .border(4.dp, color = Persian_indigo),
+
+        ) {
+        Text(
+            textAlign = TextAlign.Center,
+            text = "Sign up",
+            fontSize = 40.sp,
+            color = Color.Black,
+        )
+    }
+}
+
+@Composable
+fun lines(){
+    Row(modifier = Modifier
+        .fillMaxWidth(),
+        horizontalArrangement = Arrangement.Center,
+    ){
+        Box(
+            modifier = Modifier
+                .padding(top = 20.dp, end = 10.dp)
+                .height(height = 3.dp)
+                .width(260.dp)
+                .background(color = Color.Black)
+        )
+
+        Text(
+            textAlign = TextAlign.Center,
+            text = "OR",
+            fontSize = 30.sp,
+            color = Color.Black,
+        )
+
+        Box(
+            modifier = Modifier
+                .padding(top = 20.dp, start = 10.dp)
+                .width(260.dp)
+                .height(height = 3.dp)
+                .background(color = Color.Black)
+        )
+    }
+}
